cdk/requirements.txt
aws-cdk.aws-ec2==1.85.0
aws-cdk.aws-iam==1.85.0
aws-cdk.aws-lambda==1.85.0
aws-cdk.aws-logs==1.85.0
aws-cdk.aws-sagemaker==1.85.0
aws-cdk.core==1.85.0
